Isometric Tilemap Setup
The camera Custom Axis Sorting has been changed to 0, 1,-0.26 for sorting solutions.

The Tilemap mode is changing from Chunk to Individual. We have to do this if we want inlayed terrain or stacking of any kind to be easily implimented. It really shouldnt be too more demanding on the system to do it this way, but the example "water trench" I made here relies on this to work correctly.

The grid, and all subsequent tile palettes we make need to be in Isometric Z as Y. This may break our test tile set we downloaded off the asset store. I'm going to make basic terrain blocks today so hopfully this doesnt break too long.